moving tests out of consensus.rs
